Move contour function to contour.py
