+++ pygmt/src/__init__.py
@@ -5,6 +5,7 @@ Source code for PyGMT modules.
 from pygmt.src.basemap import basemap
 from pygmt.src.coast import coast
 from pygmt.src.colorbar import colorbar
+from pygmt.src.contour import contour
 from pygmt.src.grdcontour import grdcontour
 from pygmt.src.image import image
 from pygmt.src.legend import legend
